Minor improvements to BadZipFile and BadZipfile docs.
I made the doc for the compat alias BadZipfile shorter and used a
directive to document deprecation.

I figured there was no point of talking about zipfile.error (“the old
name” that’s older than the other old name BadZipfile) in the 3.x docs
so I just removed it.

--HG--
branch : 3.2