Add a hint that CSD == Service Pack.
People searcing for the way to get a "service pack" will never find that we
provide it here, and people that find this function won't know what CSD is
until they run the function. On top of this, they won't know what the value
means unless they really have a service pack installed.

CSD, or Customer Service Diagnostics, is apparently no longer used, and was
rarely used term at that. Most references to it online are from
universities making Windows 2000 and XP service packs available to students.

--HG--
branch : 3.2