Issue #15663: OS X installer builtin Tcl/Tk support
Make it easier for users to make use of the backup _tkinter linked
with the third-party Tcl and Tk frameworks in /Library/Frameworks.
The two tkinter variants are now installed in separate directories
under a new lib-tkinter.  This allows per-user selection by
manipulating sys.path, directly or with PYTHONPATH.  If this
proves useful, we can supply a more convenient user interface
to supply the paths.  For now, this remains somewhat experimental.

--HG--
branch : 3.3